nest module template in node template.
